+++ curso-spring-restapi/src/main/java/com/jpbandeira/springrestapi/services/CategoriaService.java
@@ -15,7 +15,7 @@ public class CategoriaService {
 	@Autowired
 	private CategoriaRepository categoriaRepository;
 
-	public Categoria buscarCategoria(Integer id) throws ObjectNotFoundException {
+	public Categoria buscarCategoria(Long id) throws ObjectNotFoundException {
 		Optional<Categoria> objetoCategoria = categoriaRepository.findById(id);
 		return objetoCategoria.orElseThrow(() -> new ObjectNotFoundException("Objeto não encontrado! Id: " + id + ", Tipo: " + Categoria.class.getName()));
 	}
